Add back a "Link to Sound"
Felt bad about removing functionality, so add it back in an alternate
form. It is just a link that gets updated every time the sound
changes.  A more modern way would be to do this with `#` which can be
updated in place instead of page parameters, but for now this gets the
job done.